requirements.txt
tqdm
colorama
pyfiglet
prettytable
pytest-cov